requirements.txt
numpy
pandas
matplotlib
pycurl
requests
pyyaml
networkx
